+++ CHANGELOG.rst
@@ -41,6 +41,9 @@ Deprecated
 - ``Path.__call__`` is deprecated and will be removed in v5.0.0. Use the
   ``absolute`` or ``relative`` properties instead (`#794
   <https://github.com/omni-us/jsonargparse/pull/794>`__).
+- ``strip_meta`` is deprecated and will be removed in v5.0.0. Instead use
+  ``.clone(with_meta=False)`` (`#795
+  <https://github.com/omni-us/jsonargparse/pull/795>`__).
 
 
 v4.42.0 (2025-10-14)